add parameters to signIn()
Sign in can be to more than one endpoint with their own cognito client ID and their own redirect landing location based on one of several roles. It means using some prop drilling to get the parameters all the way to the right function, but it seems like a decent compromise.