load parameter configs from api into a react context
and make the context available to the entire app. Because of the asyncronous external call, any component that uses the appConfig values will need to add it to its useEffect dependencies so it knows when the app config is available. Or it needs to be the child of a component that does this. It's only relevant if that page is the first thing loaded, subsequent page loads will already have the appConfig from the initial page load.