patch role home pages for post signup
if there is an existing login cookie. This change moves the post signup conditional and pre-emptively removes any existing login cookie that may already be there. Trying to sign up when you are already logged in as a different role is not likely to happen in production, but it is pretty easy to wind up doing if you are testing.