Adds signature to registration data
Ensures the registration signature is correctly included when submitting registration data, both for initial account creation and amendment scenarios.

Updates the support professional registration form to include the signature.

This means the Authorized Individual essentially registers twice, because the signature is on a different screen from  the rest of the data entry, but that seems like an acceptable compromise.