text change that doesn't even really make sense
The lookback period isn't displayed in the form at all, it is only in the invitation email. So this version of the text makes even less sense than what it was before, and both versions are basically wrong as the information is just not there however one refers to it. This should just be changed to "see the lookback period in the invitation email"