add robust validation for the formConstraint url segment
because this is coming directly from the browser and could be anything.  Once we validate and parse it here, we don't need to validated or alter it in any of the child components.

This also removes an issue where the formContraint parameter was getting mutated, which isn't that big a deal but just happens to be a no-no in React and generates warnings.